separating forward and backward phase

+++ src/backward_phase.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <mpi.h>
 
-void backward_phase(int rank, int* ids, int next_id, int block_size, int m, int n, int h, int V, int lr, double wd, double* local_gradient_La, double* local_gradient_Lo, double* C, double* x, double* H, double* o, double* d, double* local_U, double* local_gradient_Ly, double* gradient_Lx, double* gradient_La, double* local_b, double* local_p) {
+void backward_phase(int rank, int* ids, int next_id, int block_size, int m, int n, int h, int V, double lr, double wd, double* local_gradient_La, double* local_gradient_Lo, double* C, double* x, double* H, double* o, double* d, double* local_U, double* local_gradient_Ly, double* gradient_Lx, double* gradient_La, double* local_b, double* local_p) {
     // perform backward gradient for output units in i-th block
             for (int i = 0; i < h; i++) {
                 local_gradient_La[i] = 0.0;
